requirements.txt
selenium
undetected-chromedriver
setuptools
python-dotenv
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
glfw
numpy
